remove v1 files and move v3 files up a level, fix references
