+++ src/components/fields/schemaFields/SchemaField.stories.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2021 PixieBrix, Inc.
+ * Copyright (C) 2022 PixieBrix, Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -25,7 +25,7 @@ import { getFieldNamesFromPathString } from "@/runtime/pathHelpers";
 import { action } from "@storybook/addon-actions";
 
 export default {
-  title: "Fields/v3/SchemaField",
+  title: "Fields/SchemaField",
   component: SchemaField,
 } as ComponentMeta<typeof SchemaField>;
