+++ src/devTools/editor/tabs/editTab/editorNodeConfigPanel/EditorNodeConfigPanel.tsx
@@ -19,8 +19,8 @@ import React, { useMemo } from "react";
 import { Button, Col, Row } from "react-bootstrap";
 import { RegistryId } from "@/core";
 import ConnectedFieldTemplate from "@/components/form/ConnectedFieldTemplate";
-import BlockConfigurationV1 from "@/devTools/editor/tabs/effect/v1/BlockConfiguration";
-import BlockConfigurationV3 from "@/devTools/editor/tabs/effect/v3/BlockConfiguration";
+import BlockConfigurationV1 from "@/devTools/editor/tabs/effect/BlockConfiguration";
+import BlockConfigurationV3 from "@/devTools/editor/tabs/effect/BlockConfiguration";
 import { useAsyncState } from "@/hooks/common";
 import blockRegistry from "@/blocks/registry";
 import { getType } from "@/blocks/util";
