+++ src/devTools/editor/tabs/effect/BlockConfiguration.test.tsx
